+++ packages/control-property-editor/src/middleware.ts
@@ -21,7 +21,9 @@ import {
     toggleExpandMessage,
     readMessage,
     expandableMessage,
-    toggleModalMessage
+    toggleModalMessage,
+    requestControlContextMenu,
+    executeContextMenuAction
 } from '@sap-ux-private/control-property-editor-common';
 
 import type reducer from './slice';
@@ -82,7 +84,9 @@ export const communicationMiddleware: Middleware<Dispatch<ExternalAction>, Retur
                 case save.type:
                 case selectControl.type:
                 case addExtensionPoint.type:
-                case externalFileChange.type: {
+                case externalFileChange.type:
+                case executeContextMenuAction.type:
+                case requestControlContextMenu.pending.type: {
                     sendAction(action);
                     break;
                 }