+++ packages/control-property-editor/src/panels/outline/Tree.tsx
@@ -3,11 +3,24 @@ import type { ReactElement } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import type { IGroup, IGroupRenderProps, IGroupHeaderProps } from '@fluentui/react';
 import { Icon } from '@fluentui/react';
-import { UIList, UiIcons } from '@sap-ux/ui-components';
+import {
+    UIContextualMenu,
+    UIContextualMenuLayoutType,
+    UIDirectionalHint,
+    UIList,
+    UiIcons
+} from '@sap-ux/ui-components';
+import type { UIContextualMenuItem } from '@sap-ux/ui-components';
 import { useTranslation } from 'react-i18next';
 
-import { selectControl, reportTelemetry, addExtensionPoint } from '@sap-ux-private/control-property-editor-common';
-import type { Control, OutlineNode } from '@sap-ux-private/control-property-editor-common';
+import {
+    selectControl,
+    reportTelemetry,
+    addExtensionPoint,
+    executeContextMenuAction,
+    requestControlContextMenu
+} from '@sap-ux-private/control-property-editor-common';
+import type { ContextMenu, Control, OutlineNode } from '@sap-ux-private/control-property-editor-common';
 
 import type { RootState } from '../../store';
 import type { ControlChanges, FilterOptions } from '../../slice';
@@ -39,11 +52,13 @@ export const Tree = (): ReactElement => {
         group: undefined,
         cell: undefined
     });
-
+    const [showActionContextualMenu, setShowActionContextualMenu] = useState<OutlineNodeItem | undefined>();
     const filterQuery = useSelector<RootState, FilterOptions[]>((state) => state.filterQuery);
     const selectedControl = useSelector<RootState, Control | undefined>((state) => state.selectedControl);
     const controlChanges = useSelector<RootState, ControlChanges>((state) => state.changes.controls);
+    const contextMenu = useSelector<RootState, ContextMenu | undefined>((state) => state.contextMenu);
     const model: OutlineNode[] = useSelector<RootState, OutlineNode[]>((state) => state.outline);
+    const isNavigationMode = useSelector<RootState, boolean>((state) => state.appMode === 'navigation');
 
     const { groups, items } = useMemo(() => {
         const items: OutlineNodeItem[] = [];
@@ -53,9 +68,6 @@ export const Tree = (): ReactElement => {
 
     const selectedClassName =
         localStorage.getItem('theme') === 'high contrast' ? 'app-panel-hc-selected-bg' : 'app-panel-selected-bg';
-
-    const tooltipEventListeners: Record<string, (event: MouseEvent) => void> = {};
-    let currentOpenTooltipId: string | null = null;
 
     useEffect(() => {
         if (selection.cell === undefined && selection.group === undefined && selectedControl !== undefined) {
@@ -92,72 +104,6 @@ export const Tree = (): ReactElement => {
             }, 0);
         }
     }, []);
-
-    /**
-     * Closes a tooltip and removes the associated event listener.
-     *
-     * @param tooltipId The unique identifier for the tooltip.
-     * @param eventListener The specific event listener to remove (optional).
-     */
-    const closeTooltip = (tooltipId: string, eventListener?: (event: MouseEvent) => void) => {
-        const tooltip = document.getElementById(tooltipId);
-
-        if (tooltip) {
-            tooltip.style.visibility = 'hidden';
-            tooltip.style.opacity = '0';
-        }
-
-        if (eventListener) {
-            document.removeEventListener('click', eventListener);
-        } else if (tooltipId in tooltipEventListeners) {
-            document.removeEventListener('click', tooltipEventListeners[tooltipId]);
-            delete tooltipEventListeners[tooltipId];
-        }
-
-        if (currentOpenTooltipId === tooltipId) {
-            currentOpenTooltipId = null;
-        }
-    };
-
-    /**
-     * Handles the opening of a tooltip and associates an event listener with it.
-     *
-     * @param e The click event that triggered the tooltip.
-     * @param tooltipId The unique identifier for the tooltip.
-     */
-    const handleOpenTooltip = (e: React.MouseEvent<HTMLSpanElement, MouseEvent>, tooltipId: string) => {
-        e.preventDefault();
-        const tooltip = document.getElementById(tooltipId);
-
-        if (currentOpenTooltipId && currentOpenTooltipId !== tooltipId) {
-            closeTooltip(currentOpenTooltipId);
-        }
-
-        if (tooltip) {
-            tooltip.style.visibility = 'visible';
-            tooltip.style.opacity = '1';
-            currentOpenTooltipId = tooltipId;
-
-            const handleCloseTooltip = (event: MouseEvent) => {
-                if (!tooltip.contains(event.target as Node)) {
-                    closeTooltip(tooltipId, handleCloseTooltip);
-                }
-            };
-
-            tooltipEventListeners[tooltipId] = handleCloseTooltip;
-
-            document.addEventListener('click', handleCloseTooltip);
-        } else {
-            console.warn(`Tooltip with id ${tooltipId} not found`);
-        }
-    };
-
-    const handleOpenFragmentDialog = (data: OutlineNode, tooltipId: string) => {
-        if (data.controlType === 'sap.ui.extensionpoint') {
-            closeTooltip(tooltipId);
-            dispatch(addExtensionPoint(data));
-        }
-    };
 
     /**
      * Find in group.
@@ -256,7 +202,34 @@ export const Tree = (): ReactElement => {
     if (items.length === 0 && groups.length === 0) {
         return <NoControlFound />;
     }
-
+    const onContextMenuAction = (
+        e: React.MouseEvent<HTMLSpanElement, MouseEvent>,
+        cellItem?: OutlineNodeItem,
+        headerItem?: IGroup
+    ): void => {
+        e.preventDefault();
+        let selectAction;
+        const item = cellItem ?? headerItem?.data;
+        const controlId = item.controlId;
+        if (headerItem) {
+            setSelection({
+                group: headerItem,
+                cell: undefined
+            });
+            selectAction = selectControl(headerItem.key);
+        } else {
+            setSelection({
+                group: undefined,
+                cell: item
+            });
+            selectAction = selectControl(controlId);
+        }
+        dispatch(selectAction);
+        if (item?.controlType !== 'sap.ui.extensionpoint') {
+            dispatch(requestControlContextMenu.pending(controlId));
+        }
+        setShowActionContextualMenu(item);
+    };
     const onSelectCell = (item: OutlineNodeItem): void => {
         setSelection({
             group: undefined,
@@ -280,6 +253,49 @@ export const Tree = (): ReactElement => {
             const action = selectControl(node.key);
             dispatch(action);
         }
+    };
+    /**
+     * Build menu items for context menu.
+     *
+     * @returns ReactElement
+     */
+    const buildMenuItems = function (): UIContextualMenuItem[] {
+        if (!showActionContextualMenu) {
+            return [];
+        }
+        const { controlId, controlType } = showActionContextualMenu;
+        const isExtensionPoint = controlType === 'sap.ui.extensionpoint';
+        const children = isExtensionPoint
+            ? [
+                  {
+                      id: '',
+                      title: t('ADD_FRAGMENT_AT_EXTENSION_POINT'),
+                      enabled: true
+                  }
+              ]
+            : contextMenu?.contextMenuItems;
+        return (children ?? []).map((child, index) => {
+            return {
+                key: `${controlId}-${child.id}-${index}`,
+                text: child?.title,
+                disabled: isNavigationMode ? true : !child.enabled,
+                title: isNavigationMode
+                    ? t('CONTEXT_MENU_ACTION_DISABLED_IN_NAVIGATION_MODE')
+                    : child?.tooltip ?? child?.title,
+                onClick(): void {
+                    if (isExtensionPoint) {
+                        dispatch(addExtensionPoint(showActionContextualMenu));
+                    } else {
+                        dispatch(
+                            executeContextMenuAction({
+                                controlId,
+                                actionName: child.id
+                            })
+                        );
+                    }
+                }
+            };
+        });
     };
     const onRenderCell = (nestingDepth?: number, item?: OutlineNodeItem, itemIndex?: number): React.ReactNode => {
         const paddingValue = (item?.level ?? 0) * 10 + 45;
@@ -307,8 +323,6 @@ export const Tree = (): ReactElement => {
         const isExtensionPoint = item?.controlType === 'sap.ui.extensionpoint';
         const hasDefaultContent = item?.hasDefaultContent || false;
 
-        const tooltipId = `tooltip--${item?.name}`;
-
         const cellName = hasDefaultContent
             ? t('EXTENSION_POINT_HAS_DEFAULT_CONTENT_TEXT', { name: item?.name })
             : item?.name;
@@ -316,14 +330,11 @@ export const Tree = (): ReactElement => {
             <div
                 aria-hidden
                 id={item.controlId}
+                data-control-id={isExtensionPoint ? `${item.controlId}--extensionPoint` : item.controlId}
                 className={classNames.join(' ')}
-                onClick={(): void => onSelectCell(item)}>
-                <span
-                    {...props}
-                    data-testid={isExtensionPoint ? 'tooltip-container' : ''}
-                    style={{ paddingLeft: paddingValue }}
-                    className={`tree-cell ${isExtensionPoint ? 'tooltip-container' : ''}`}
-                    onContextMenu={(e) => isExtensionPoint && handleOpenTooltip(e, tooltipId)}>
+                onClick={(): void => onSelectCell(item)}
+                onContextMenu={(e) => onContextMenuAction(e, item)}>
+                <span {...props} style={{ paddingLeft: paddingValue }} className="tree-cell">
                     {isExtensionPoint && <Icon className="extension-icon" iconName={UiIcons.DataSource} />}
 
                     <div
@@ -336,18 +347,7 @@ export const Tree = (): ReactElement => {
                         title={isExtensionPoint ? item?.name : ''}>
                         {cellName}
                     </div>
-
-                    {isExtensionPoint && (
-                        <div id={tooltipId} className="tooltip">
-                            <button
-                                data-testid="tooltip-dialog-button"
-                                onClick={() => handleOpenFragmentDialog(item, tooltipId)}>
-                                {t('ADD_FRAGMENT_AT_EXTENSION_POINT')}
-                            </button>
-                        </div>
-                    )}
                 </span>
-
                 <div style={{ marginLeft: '10px', marginRight: '10px' }}>{indicator}</div>
             </div>
         ) : null;
@@ -405,24 +405,21 @@ export const Tree = (): ReactElement => {
             <></>
         );
 
-        const tooltipId = `tooltip--${groupName}`;
         const headerName = hasDefaultContent
             ? t('EXTENSION_POINT_HAS_DEFAULT_CONTENT_TEXT', { name: groupName })
             : groupName;
         return (
             <div
+                data-control-id={isExtensionPoint ? `${data.controlId}--extensionPoint` : data.controlId}
                 {...refProps}
                 aria-hidden
                 className={`${selectNode} tree-row ${focusEditable}`}
-                onClick={(): void => onSelectHeader(groupHeaderProps?.group)}>
-                <span
-                    style={{ paddingLeft: paddingValue }}
-                    data-testid="tooltip-container"
-                    className={`tree-cell ${isExtensionPoint ? 'tooltip-container' : ''}`}
-                    onContextMenu={(e) => isExtensionPoint && handleOpenTooltip(e, tooltipId)}>
+                onClick={(): void => onSelectHeader(groupHeaderProps?.group)}
+                onContextMenu={(e) => onContextMenuAction(e, undefined, groupHeaderProps?.group)}>
+                <span style={{ paddingLeft: paddingValue }} className="tree-cell">
                     {groupHeaderProps?.group?.count !== 0 && (
                         <Icon
-                            className={`${chevronTransform}`}
+                            className={chevronTransform}
                             iconName={UiIcons.Chevron}
                             onClick={(event) => {
                                 onToggleCollapse(groupHeaderProps);
@@ -442,16 +439,6 @@ export const Tree = (): ReactElement => {
                         title={isExtensionPoint ? groupName : ''}>
                         {headerName}
                     </div>
-
-                    {isExtensionPoint && (
-                        <div id={tooltipId} className="tooltip">
-                            <button
-                                data-testid="tooltip-dialog-button"
-                                onClick={() => handleOpenFragmentDialog(groupHeaderProps?.group?.data, tooltipId)}>
-                                {t('ADD_FRAGMENT_AT_EXTENSION_POINT')}
-                            </button>
-                        </div>
-                    )}
                 </span>
                 <div style={{ marginLeft: '10px', marginRight: '10px' }}>{indicator}</div>
             </div>
@@ -482,6 +469,22 @@ export const Tree = (): ReactElement => {
                 onSelect={onSelectHeader}
                 groupProps={groupRenderProps}
             />
+            {showActionContextualMenu && (
+                <UIContextualMenu
+                    layoutType={UIContextualMenuLayoutType.ContextualMenu}
+                    showSubmenuBeneath={true}
+                    target={`[data-control-id="${
+                        showActionContextualMenu.controlType === 'sap.ui.extensionpoint'
+                            ? `${showActionContextualMenu.controlId}--extensionPoint`
+                            : showActionContextualMenu.controlId
+                    }"]`}
+                    isBeakVisible={true}
+                    items={buildMenuItems()}
+                    directionalHint={UIDirectionalHint.bottomRightEdge}
+                    onDismiss={() => setShowActionContextualMenu(undefined)}
+                    iconToLeft={true}
+                />
+            )}
         </div>
     );
 };