+++ packages/abap-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,98 @@
 # @sap-ux/abap-deploy-config-sub-generator
 
+## 0.0.31
+
+### Patch Changes
+
+-   Updated dependencies [ed8a9b9]
+    -   @sap-ux/ui5-config@0.26.2
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.13
+    -   @sap-ux/abap-deploy-config-writer@0.0.86
+    -   @sap-ux/project-access@1.29.6
+    -   @sap-ux/telemetry@0.5.60
+    -   @sap-ux/fiori-generator-shared@0.7.28
+    -   @sap-ux/deploy-config-generator-shared@0.0.21
+
+## 0.0.30
+
+### Patch Changes
+
+-   Updated dependencies [c08e1e3]
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.12
+
+## 0.0.29
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.5
+-   @sap-ux/abap-deploy-config-writer@0.0.85
+-   @sap-ux/fiori-generator-shared@0.7.27
+-   @sap-ux/telemetry@0.5.59
+-   @sap-ux/abap-deploy-config-inquirer@1.2.11
+-   @sap-ux/deploy-config-generator-shared@0.0.20
+
+## 0.0.28
+
+### Patch Changes
+
+-   Updated dependencies [65f15d9]
+    -   @sap-ux/btp-utils@1.0.1
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.10
+    -   @sap-ux/deploy-config-generator-shared@0.0.19
+    -   @sap-ux/fiori-generator-shared@0.7.26
+    -   @sap-ux/telemetry@0.5.58
+    -   @sap-ux/abap-deploy-config-writer@0.0.84
+
+## 0.0.27
+
+### Patch Changes
+
+-   9b755fc: adds new cf generator
+-   Updated dependencies [9b755fc]
+    -   @sap-ux/deploy-config-generator-shared@0.0.18
+
+## 0.0.26
+
+### Patch Changes
+
+-   Updated dependencies [29abc73]
+    -   @sap-ux/project-access@1.29.4
+    -   @sap-ux/abap-deploy-config-writer@0.0.83
+    -   @sap-ux/fiori-generator-shared@0.7.25
+    -   @sap-ux/telemetry@0.5.57
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.9
+    -   @sap-ux/deploy-config-generator-shared@0.0.17
+
+## 0.0.25
+
+### Patch Changes
+
+-   Updated dependencies [9980073]
+    -   @sap-ux/btp-utils@1.0.0
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.8
+    -   @sap-ux/deploy-config-generator-shared@0.0.16
+    -   @sap-ux/fiori-generator-shared@0.7.24
+    -   @sap-ux/telemetry@0.5.56
+    -   @sap-ux/abap-deploy-config-writer@0.0.82
+
+## 0.0.24
+
+### Patch Changes
+
+-   Updated dependencies [096b021]
+    -   @sap-ux/project-access@1.29.3
+    -   @sap-ux/abap-deploy-config-writer@0.0.81
+    -   @sap-ux/fiori-generator-shared@0.7.23
+    -   @sap-ux/telemetry@0.5.55
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.7
+    -   @sap-ux/deploy-config-generator-shared@0.0.15
+
+## 0.0.23
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.2.6
+
 ## 0.0.22
 
 ### Patch Changes