+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.33.3
+
+### Patch Changes
+
+-   @sap-ux/fiori-annotation-api@0.4.7
+
+## 0.33.2
+
+### Patch Changes
+
+-   Updated dependencies [9e5c80f]
+    -   @sap-ux/fiori-annotation-api@0.4.6
+
+## 0.33.1
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.6
+-   @sap-ux/fiori-annotation-api@0.4.5
+
+## 0.33.0
+
+### Minor Changes
+
+-   aaaf046: update validateBasePath to use dynamic webapp path from manifest.json instead of hardcoded value
+
+## 0.32.4
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.5
+-   @sap-ux/fiori-annotation-api@0.4.4
+
+## 0.32.3
+
+### Patch Changes
+
+-   Updated dependencies [29abc73]
+    -   @sap-ux/project-access@1.29.4
+    -   @sap-ux/fiori-annotation-api@0.4.3
+
+## 0.32.2
+
+### Patch Changes
+
+-   Updated dependencies [1f0bb25]
+    -   @sap-ux/fiori-annotation-api@0.4.2
+
+## 0.32.1
+
+### Patch Changes
+
+-   Updated dependencies [096b021]
+    -   @sap-ux/project-access@1.29.3
+    -   @sap-ux/fiori-annotation-api@0.4.1
+
 ## 0.32.0
 
 ### Minor Changes