+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.9.15
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.6
+-   @sap-ux/inquirer-common@0.6.15
+
+## 0.9.14
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.5
+-   @sap-ux/inquirer-common@0.6.14
+
+## 0.9.13
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.13
+
+## 0.9.12
+
+### Patch Changes
+
+-   Updated dependencies [29abc73]
+    -   @sap-ux/project-access@1.29.4
+    -   @sap-ux/inquirer-common@0.6.12
+
+## 0.9.11
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.11
+
+## 0.9.10
+
+### Patch Changes
+
+-   Updated dependencies [096b021]
+    -   @sap-ux/project-access@1.29.3
+    -   @sap-ux/inquirer-common@0.6.10
+
+## 0.9.9
+
+### Patch Changes
+
+-   Updated dependencies [5e3a5f8]
+    -   @sap-ux/inquirer-common@0.6.9
+
 ## 0.9.8
 
 ### Patch Changes