+++ packages/ui5-library-inquirer/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @sap-ux/ui5-library-inquirer
 
+## 0.3.44
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.15
+
+## 0.3.43
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.14
+
+## 0.3.42
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.13
+
+## 0.3.41
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.12
+
+## 0.3.40
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.11
+
+## 0.3.39
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.10
+
+## 0.3.38
+
+### Patch Changes
+
+-   Updated dependencies [5e3a5f8]
+    -   @sap-ux/inquirer-common@0.6.9
+
 ## 0.3.37
 
 ### Patch Changes